pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "krx-clearlog"
version = "0.1.0"
description = "Parser and Streamlit viewer for KRX clearing-and-settlement (청산결제) log messages."
readme = "README.md"
requires-python = ">=3.11,<3.12"
authors = [{ name = "Koscom" }]
license = { text = "Proprietary" }
dependencies = [
    "streamlit>=1.32,<2.0",
    "streamlit-ace>=0.1.1,<1.0",
    "sqlalchemy>=2.0,<3.0",
    "alembic>=1.13,<2.0",
    "pydantic>=2.6,<3.0",
    "pydantic-settings>=2.2,<3.0",
    "pyyaml>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "ruff>=0.4,<1.0",
]

[tool.setuptools.packages.find]
include = ["krx_parser*", "app*"]

[tool.setuptools.package-data]
krx_parser = ["schemas/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
